% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_df}
\alias{check_df}
\title{Check input data.frame}
\usage{
check_df(dat, call = rlang::caller_env())
}
\arguments{
\item{dat}{data frame input}

\item{call}{calling environment}
}
\description{
The input data frame should have a limited number of columns and at least two rows
}
