% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{create_ranges}
\alias{create_ranges}
\title{Create a range data frame}
\usage{
create_ranges()
}
\description{
This function creates an example data frame of mins and maxs for brew
  parameter settings. That is, the range of grinder settings I want to search is from 4
  to 14, temperatures from 170 to 210F, and bloom times from 0 to 60s.
}
