% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyingforacup-package.R
\docType{package}
\name{dyingforacup}
\alias{dyingforacup-package}
\alias{dyingforacup}
\title{dyingforacup: Example Package With Instantiate}
\description{
A package using `instantiate` for coffee optimization.
}
\concept{help}
\keyword{internal}
