% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_gp}
\alias{run_gp}
\title{Run the GP}
\usage{
run_gp(
  dat,
  ...,
  max_grid_size = 2000,
  param_ranges = create_ranges(),
  param_grid = NULL
)
}
\arguments{
\item{dat}{data frame input of brew parameters and rating}

\item{...}{arguments passed to cmdstanr's sample method}

\item{max_grid_size}{maximum number of grid points to evaluate}

\item{param_ranges}{upper and lower limits of parameter ranges to evaluate (ignored if param_grid is specified directly)}

\item{param_grid}{user-specified grid of brew parameters to evaluate GP at}
}
\description{
Run the GP
}
\details{
The function \code{\link[dyingforacup:create_ranges]{dyingforacup::create_ranges()}} will create an example range df.
}
