% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_gp_model}
\alias{run_gp_model}
\title{Fit the GP model.}
\usage{
run_gp_model(X, y, X_pred, ..., verbose)
}
\arguments{
\item{X}{a matrix of scaled brew parameters}

\item{y}{Numeric vector of ratings (NORMAL scale, not 0-10!)}

\item{X_pred}{a matrix of grid points of in scaled brew parameter space to evaluate the GP at}

\item{...}{Named arguments to the `sample()` method of CmdStan model}

\item{verbose}{logical indicating whether to print messages
objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>}
}
\value{
a draws data frame
}
\description{
Fit the GP Stan model and return posterior summaries.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  run_gp_model(y = rnorm(5))
}
}
\concept{models}
