% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{suggest_next}
\alias{suggest_next}
\title{Suggest the next point to try}
\usage{
suggest_next(
  dat,
  ...,
  max_grid_size = 2000,
  param_ranges = create_ranges(),
  param_grid = NULL,
  offset = 0.25
)
}
\arguments{
\item{dat}{data frame input of brew parameters and rating}

\item{...}{arguments passed to cmdstanr's sample method}

\item{max_grid_size}{maximum number of grid points to evaluate}

\item{param_ranges}{upper and lower limits of parameter ranges to evaluate (ignored if param_grid is specified directly)}

\item{param_grid}{user-specified grid of brew parameters to evaluate GP at}

\item{offset}{expected improvement hyperparameter. Higher values encourage more
exploration. Interpreted on the same scale as ratings.}
}
\description{
Suggest the next point to try
}
