% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{suggest_next}
\alias{suggest_next}
\title{Suggest the next point to try}
\usage{
suggest_next(
  dat,
  ...,
  max_grid_size = 2000,
  param_ranges = create_ranges(),
  param_grid = NULL,
  offset = 0.25,
  lambda = 0.1
)
}
\arguments{
\item{dat}{data frame input of brew parameters and rating}

\item{...}{arguments passed to cmdstanr's sample method}

\item{max_grid_size}{maximum number of grid points to evaluate}

\item{param_ranges}{upper and lower limits of parameter ranges to evaluate (ignored if param_grid is specified directly)}

\item{param_grid}{user-specified grid of brew parameters to evaluate GP at}

\item{offset}{expected improvement hyperparameter. Higher values encourage more
exploration. Interpreted on the same scale as ratings.}

\item{lambda}{tradeoff between weighting posterior predictive variance and expected
improvement at grid values.}
}
\value{
a list with elements:
\itemize{
  \item{draws_df}{a draws data frame of model parameters and grid point predictive draws f_star}
  \item{acq_df}{a data table of acquisition function values and corresponding grid point values}
  \item{suggested}{the row of \code{acq_df} that maximizes the acquisition function}
}
}
\description{
Suggest the next point to try
}
\details{
The acquisition function is \code{lambda*f_star_var + (1-lambda)*exp_imp}.
Higher values of lambda up-weight posterior predictive variance, leading to more
exploration over exploitation. Lower lambda values up-weight expected improvement over \code{max(dat$rating) - offset}.
}
